+++ crates/cubecl-core/src/frontend/element/cube_elem.rs
@@ -69,27 +69,27 @@ pub trait CubePrimitive:
         client.properties().features.type_usage(elem)
     }
 
-    fn type_size() -> u32 {
-        Self::as_type_native_unchecked().size() as u32
+    fn type_size() -> usize {
+        Self::as_type_native_unchecked().size()
     }
 
-    fn type_size_bits() -> u32 {
-        Self::as_type_native_unchecked().size_bits() as u32
+    fn type_size_bits() -> usize {
+        Self::as_type_native_unchecked().size_bits()
     }
 
-    fn packing_factor() -> u32 {
+    fn packing_factor() -> usize {
         Self::as_type_native_unchecked().packing_factor()
     }
 
-    fn __expand_type_size(scope: &Scope) -> u32 {
-        Self::as_type(scope).size() as u32
+    fn __expand_type_size(scope: &Scope) -> usize {
+        Self::as_type(scope).size()
     }
 
-    fn __expand_type_size_bits(scope: &Scope) -> u32 {
-        Self::as_type(scope).size_bits() as u32
+    fn __expand_type_size_bits(scope: &Scope) -> usize {
+        Self::as_type(scope).size_bits()
     }
 
-    fn __expand_packing_factor(scope: &Scope) -> u32 {
+    fn __expand_packing_factor(scope: &Scope) -> usize {
         Self::as_type(scope).packing_factor()
     }
 }
