+++ crates/cubecl-core/src/runtime_tests/numeric.rs
@@ -4,7 +4,7 @@ use cubecl_ir::{ElemType, FloatKind, UIntKind};
 
 #[cube(launch)]
 pub fn kernel_define<N: Numeric>(array: &mut Array<N>, #[define(N)] _elem: ElemType) {
-    array[UNIT_POS] += N::cast_from(5.0f32);
+    array[UNIT_POS as usize] += N::cast_from(5.0f32);
 }
 
 #[cube(launch)]
@@ -13,7 +13,7 @@ pub fn kernel_define_many<N: Numeric, N2: Numeric>(
     second: Array<N2>,
     #[define(N, N2)] _defines: [ElemType; 2],
 ) {
-    array[UNIT_POS] += N::cast_from(second[UNIT_POS]);
+    array[UNIT_POS as usize] += N::cast_from(second[UNIT_POS as usize]);
 }
 
 pub fn test_kernel_define<R: Runtime>(client: ComputeClient<R>) {
