+++ crates/cubecl-core/src/runtime_tests/plane.rs
@@ -7,7 +7,7 @@ use cubecl_runtime::Plane;
 
 #[cube(launch)]
 pub fn kernel_sum<F: Float>(output: &mut Tensor<F>) {
-    let val = output[UNIT_POS];
+    let val = output[UNIT_POS as usize];
     let val2 = plane_sum(val);
 
     if UNIT_POS == 0 {
@@ -17,23 +17,23 @@ pub fn kernel_sum<F: Float>(output: &mut Tensor<F>) {
 
 #[cube(launch)]
 pub fn kernel_inclusive_sum<F: Float>(output: &mut Tensor<F>) {
-    let val = output[UNIT_POS];
+    let val = output[UNIT_POS as usize];
     let val2 = plane_inclusive_sum(val);
 
-    output[UNIT_POS] = val2;
+    output[UNIT_POS as usize] = val2;
 }
 
 #[cube(launch)]
 pub fn kernel_exclusive_sum<F: Float>(output: &mut Tensor<F>) {
-    let val = output[UNIT_POS];
+    let val = output[UNIT_POS as usize];
     let val2 = plane_exclusive_sum(val);
 
-    output[UNIT_POS] = val2;
+    output[UNIT_POS as usize] = val2;
 }
 
 #[cube(launch)]
 pub fn kernel_prod<F: Float>(output: &mut Tensor<F>) {
-    let val = output[UNIT_POS];
+    let val = output[UNIT_POS as usize];
     let val2 = plane_prod(val);
 
     if UNIT_POS == 0 {
@@ -43,23 +43,23 @@ pub fn kernel_prod<F: Float>(output: &mut Tensor<F>) {
 
 #[cube(launch)]
 pub fn kernel_inclusive_prod<F: Float>(output: &mut Tensor<F>) {
-    let val = output[UNIT_POS];
+    let val = output[UNIT_POS as usize];
     let val2 = plane_inclusive_prod(val);
 
-    output[UNIT_POS] = val2;
+    output[UNIT_POS as usize] = val2;
 }
 
 #[cube(launch)]
 pub fn kernel_exclusive_prod<F: Float>(output: &mut Tensor<F>) {
-    let val = output[UNIT_POS];
+    let val = output[UNIT_POS as usize];
     let val2 = plane_exclusive_prod(val);
 
-    output[UNIT_POS] = val2;
+    output[UNIT_POS as usize] = val2;
 }
 
 #[cube(launch)]
 pub fn kernel_max<F: Float>(output: &mut Tensor<F>) {
-    let val = output[UNIT_POS];
+    let val = output[UNIT_POS as usize];
     let val2 = plane_max(val);
 
     if UNIT_POS == 0 {
@@ -69,7 +69,7 @@ pub fn kernel_max<F: Float>(output: &mut Tensor<F>) {
 
 #[cube(launch)]
 pub fn kernel_min<F: Float>(output: &mut Tensor<F>) {
-    let val = output[UNIT_POS];
+    let val = output[UNIT_POS as usize];
     let val2 = plane_min(val);
 
     if UNIT_POS == 0 {
@@ -79,16 +79,16 @@ pub fn kernel_min<F: Float>(output: &mut Tensor<F>) {
 
 #[cube(launch)]
 pub fn kernel_all<F: Float>(output: &mut Tensor<F>) {
-    let val = output[UNIT_POS];
+    let val = output[UNIT_POS as usize];
     let val2 = plane_all(val < F::new(5.0));
-    output[UNIT_POS] = F::cast_from(val2);
+    output[UNIT_POS as usize] = F::cast_from(val2);
 }
 
 #[cube(launch)]
 pub fn kernel_any<F: Float>(output: &mut Tensor<F>) {
-    let val = output[UNIT_POS];
+    let val = output[UNIT_POS as usize];
     let val2 = plane_any(val > F::new(5.0));
-    output[UNIT_POS] = F::cast_from(val2);
+    output[UNIT_POS as usize] = F::cast_from(val2);
 }
 
 #[cube(launch)]
@@ -101,7 +101,7 @@ pub fn kernel_elect<F: Float>(output: &mut Tensor<F>) {
 
 #[cube(launch)]
 pub fn kernel_broadcast<F: Float>(output: &mut Tensor<F>) {
-    let val = output[UNIT_POS];
+    let val = output[UNIT_POS as usize];
     let val2 = plane_broadcast(val, 2);
 
     if UNIT_POS == 0 {
@@ -120,7 +120,7 @@ pub fn kernel_ballot(output: &mut Tensor<Line<u32>>) {
 
 #[cube(launch)]
 pub fn kernel_shuffle<F: Float>(output: &mut Tensor<F>) {
-    let val = output[UNIT_POS];
+    let val = output[UNIT_POS as usize];
     let val2 = plane_shuffle(val, 0); // All lanes read from lane 0
 
     if UNIT_POS == 0 {
@@ -130,26 +130,26 @@ pub fn kernel_shuffle<F: Float>(output: &mut Tensor<F>) {
 
 #[cube(launch)]
 pub fn kernel_shuffle_xor<F: Float>(output: &mut Tensor<F>) {
-    let val = output[UNIT_POS];
+    let val = output[UNIT_POS as usize];
     let val2 = plane_shuffle_xor(val, 1);
 
-    output[UNIT_POS] = val2;
+    output[UNIT_POS as usize] = val2;
 }
 
 #[cube(launch)]
 pub fn kernel_shuffle_up<F: Float>(output: &mut Tensor<F>) {
-    let val = output[UNIT_POS];
+    let val = output[UNIT_POS as usize];
     let val2 = plane_shuffle_up(val, 1);
 
-    output[UNIT_POS] = val2;
+    output[UNIT_POS as usize] = val2;
 }
 
 #[cube(launch)]
 pub fn kernel_shuffle_down<F: Float>(output: &mut Tensor<F>) {
-    let val = output[UNIT_POS];
+    let val = output[UNIT_POS as usize];
     let val2 = plane_shuffle_down(val, 1);
 
-    output[UNIT_POS] = val2;
+    output[UNIT_POS as usize] = val2;
 }
 
 pub fn test_plane_sum<
